Add APOLLO geometry
The APOLLO detector geometry is included according to the technical drawings.
Simplifications include absence of support rings around the crystals, the
support structure for the half-sphere and a minimum variation of its shape
in the border (cut angle).
Only CsI crystals are included at the moment.
The wrapping material thickness, the LaBr3 crystals, which position are occupied
by each crystal type and the exact displacement of each crystal along their rotation
axis (Sci_eccentricity parameter) is still to be fixed.