Update AtEventdisplay for AT-TPC and remove depricated classes
AT-TPC event display will now show upto 1000 waveforms when DrawAllPads option is on. This can be increased using the const static variable in AtEventDrawTask.h. Set the map type in AtEventDrawTask to AtTpcMap because that class is only used to display TPC data. For proto-type TPC data, AtEventDrawTaskProto should be used. Also changed headers and forward declerations in AtEventDrawTask to match the coding convensions.