Support PSA and viewer reading/writing arbitrary branches for raw events
AtPSAtask and AtEventDrawTask now allow you to specify the branch to read to get the AtRawEvent from. This allows the user to choose at runtime if they want to look at the raw or filtered data. The PSA task also allows you to set the name of the output branch so it isn't just the default "AtEventH". This could be used if you're trying multiple filters in a single run for example. This should have no effect on existing code as the default branch names are unchanged from their previously hard coded values. NB there seems to be a bug in the viewer that it will not display anything, even the pad plane unless there is a branch with type AtEvent