Add linking between HiRAEVT trees and TPC data.
Will link using the interval between events, which allows for
timestamps with a different clock frequecny to be used. If the TPC
timestamp is corrupted, it will switch to matching over the absolute
value of the timestamp with a different window specified by the user.
In the output EVT tree diagnostic graphs of the absolute difference and
relative interval difference are made for each timetamp in the `AtRawEvent`.

The linked will respect requests from other tasks to skip writing events
to disk when they `MarkFill(false)` on the run. For this to work, the linker
run has to be the last task added to the run.

A new run type was added, AtRunAna which exposes the internal variable
`kMarkFill` to tasks so the linker can work properly.