Refactor pad reference and add threshold to AGET ch0 filter
Rather than use a vector for the pad reference where one has
to remeber which element is which ID, moved to a struct where
the IDs can be indexed by name. The initialization is identical
if done with initializer lists.

The AtFilterSubtraction class will only add the ch0 of an aget to
the baseline subtraction if there was no event data there by
checking the maximum ADC value in the pad is less than some threshold.