Refactor PadReference to support std::hash
Refactored PadReference bringing it out of AtMap so the required
machinary could be added to use it as a key in the hash map storing
the electronic mapping in AtMap.

Because this was done, removed the extra hash functions created
to handle aux pads in the AtHDFParserTask. They use the same
hash functions as the std methods.