Refactor AtHDFParserTask to accept a pointer to AtMap
This allows the detector map to be created elsewhere, and passed
in. It removes the need for this class to know about what detectors
have been implemented as subclasses of AtMap and removed all of
the dynamic_casts that were floating around in this class.

Also took the time to breakup the main body into private functions
to make it more readable/maintainable.