Refactor how AtRawEvent handles its pads
Added an explicit data structure to contain the aux pads in an
event, and made AtRawEvent responsible for the creation and
memory managment of pads, both normal and aux.

The seperation of the aux pads was driven by a desire to have
direct access to them, rather than having to look through every
pad in an event to find an aux pad. They are indexed by a string
describing the pad.

Rather than creating a pad on the heap, and passing it to the event
where it is coppied, AtRawEvent will create the pad and pass back a
referece to the new pad. This makes the ownership of the pads more
explicit, and make null-ptr exceptions harder to get.