Refactor AtMCPoint, and create new AtSiPoint and cleanup AtTpc class
AtMCPoint only tracks the "input" position and momentum now.
fMomOut and fPosOut were not used for the TPC in any current code
and were only filled at the end of a track (or when it exited the
drift volume).

The example script for analyzing SiArray data does use these variables
so created AtSiPoint which inherits from AtMCpoint and adds the
removed behavior back in.

Started cleanup of AtTpc class, refactoring to make it easier to read.
This is a lead into making the tracking of the inital energy and angle
more generator agnostic