Fix memory bug in AtRawEvent
Sometimes, when the vector holding AtPads was resized when adding
a new pad, it would invalidate reference to the most recent pad added
this caused undefined behavior. To fix, switched to unique_ptr<AtPad>
whose underlying pointer address is not effected by resizing the vector
as the unique_ptr is just moved if needed.