Add initial working test suite
To run all tests run the bash script macro/tests/runAllTest.sh
To the screen it will print a summary of each test run and the
return value. The output of the test scripts run are saved in a
text file in each test directory (test.log).

Note:
* The greometry files will need to be generated before running the
tests for the first time.
* The TPC test comes with the required data
* GADGET tests must be run on fishtank (test data is too large)
* SpecMAT test requires the ROOT file TTreesGETrun_9993.root be
placed in the tests/SpecMAT/data folder.