Add GRAW unpacker (AtCore2) to the unified unpacker task.
The unpacker requires a number of cobo/asads to be set in the constructor
if this is larger than 1, then it assumes seperated data. In this case, when
setting the input data file, that data file should be a txt file containing
a list of GRAW files to load in. One on each line. To assign each file to the
correct GETDecoder, an "IDString" must also be passed in. For each GETDecoder
n \in [0, fNumDecoders) we will search through the list of files and assign the
file to the nth decoder if it contains the substring Form(IDString, n).

If the data is not seperated, then we just add the input file to the GETDecoder