Set the RUNPATH on install and build
This is a list of directories packaged into the compiled binary that are
used to seach for the dependecies of a shared library. This means the user
doesn't have to fiddle around with setting their LD_LIBRARY_PATH

Becuase we are exporing the RUNPATH rather than the older RPATH, the user
can override the search behavior by specifing a location on the
LD_LIBRARY_PATH. The search order goes:
1. RPATH
2. LD_LIBRARY_PATH
3. RUNPATH

Adds target for genfit in the FindGENFIT2 module. This allows us to
use targets to handle GENFIT's dependecies (which is all of ROOT)