Finish work on build system (baring final tests)
Now fully exporting all dependencies and the info required for a
user to find all dependencies that the code was built with.
The user can override what version of a dependency is found
by setting the cmake or enviroment variables SIMPATH, FAIRROOTPATH,
and GENFIT, otherwise we will use the values ATTPCROOT was
installed with.