Build modernization (#112)
* Add ability to clone a hit into an AtEvent

* Move more to AtTools and update README with more test info

* Work on build system.

Removed all of the old CMake files.
Removed uneeded calls to FairMacro functions, in particular
one (that set ${BASE_INCLUDE_DIRECTORIES} was making calls to
global CMake functions link_directories and include_directories.

Changed how VMC is handles to use aliased targets rather than setting
a variable which is used other places to link to VMC. Becuase FairRoot
exposes some VMC headers anything that builds against FairRoot must also
build against those headers (we don't need the libraries though) so added
those headers to the FAIRROOT_INCLUDE_DIR variable as a hacky solution
until FairRoot actually supports targets.

Move the code to print the summary to a cmake script so it doesn't
clog up the top-level cmake document

* Get ATTPCROOT installing with targets

* Will now build against more installs of HDF5

If it can't find a config file to import the required target
it will generate it based on the output of FindHDF5.cmake

* Set the RUNPATH on install and build

This is a list of directories packaged into the compiled binary that are
used to seach for the dependencies of a shared library. This means the user
doesn't have to fiddle around with setting their LD_LIBRARY_PATH

Because we are exporting the RUNPATH rather than the older RPATH, the user
can override the search behavior by specifying a location on the
LD_LIBRARY_PATH. The search order goes:
1. RPATH
2. LD_LIBRARY_PATH
3. RUNPATH

Adds target for genfit in the FindGENFIT2 module. This allows us to
use targets to handle GENFIT's dependencies (which is all of ROOT)

* Update FindROOT to first search for config

If it cannot it will default to old behavior

* Modify more Find... to manually export targets

There should be no more need to manually set the link or include
directories for external dependencies

* Build system properly exports all of the info needed to use project

* Cleanup build scripts

Export our namespace with our targets, and add an alias for internal use

Add ability to print which optional dependencies were not found which
is easier to digest then the list of what was found

* Finish work on build system (baring final tests)

Now fully exporting all dependencies and the info required for a
user to find all dependencies that the code was built with.
The user can override what version of a dependency is found
by setting the cmake or environment variables SIMPATH, FAIRROOTPATH,
and GENFIT, otherwise we will use the values ATTPCROOT was
installed with.