Add FFT Filter (#9)
* Add FFT Filter

IO seems to be working, but the backwards transformation after
frequency cuts have been applied has odd artifacts. Without frequency
cuts returns the input data after transforming back

* Change IO for FFT filter

To avoid have an abundance of copies of the same event in the same
tree there is now a single parameter for if we should save
fourier-space representation of the traces. If we should then modify
the input event to store the represetation before filtering and add
the representation to the output event.

* Add additional functions to get info complex FFT transform

* Run linters on code and cleanup