Replace PCL with AtSampleConsensus and refactor AtTrack
AtLmedsMod, AtMlesacMod, AtRansacMod replaced by
AtSampleConsensus. The different esimators exist as free functions in
namespace SampleConsensus and are used through the static class
AtEstimator. The documentation describing this is in the doxygen
module Sample Consensus.

Information on the shape of tracks (AtPattern) is in libary AtData,
namespace AtPatterns. Documentation is in the doxygen module Track
Patterns. AtPatterns are saved in AtTrack instead of fit
parameters. Currently implemented two patterns a line and a
circle. Patterns have the information to draw themselves and will be
drawn by the event viewer.

Moved random sampling out of AtRansacMod and into its own class in
AtTools. It really doesn't have anything to do with RANSAC
specificially and could, in principle, be useful in other code.

Refactor hit sampling to allow for many more options based on
arbitrary PDFs.

PCL is still around in hc code and AtTrackFinderHC

Remove isNoise from AtTrack it is tracked in AtPatternEvent instead