Fix angle reconstruction in AtPRA
The angle reconstruction assumed the direction vector from the ransac
fit was normalized to one. The model parameters from AtPatternLine did
not ensure that. Verified angles from PCL and AtSampleConsensus
match after fix.

The order models are returned in AtSampleConsensus is
how well they fit their inliers, for AtPRA what we want is the circle
with the most points (not few points that match a circle
well).

Track merging in EFitterAnalysis was always being
skipped becasue AtTrack.fIsMergered was defaulting to true instead of
false.

The ordering of the hits seems to be off. From what I can tell, they
are sorted only twice. One in AtPSAsimple2 right before an event is
saved, and once in AtFitter right before merged tracks are
reclusterized.