Update bergen macros
I just reused the code in mira.C by including it and adding a header
guard. The fucntions loadPad(padNum) and loadMesh() will return
historgrams to the given pad/mesh. If the pad doesn't exist in the
event, you will get a nullptr.

I also added some comments to you (the code probably wasn't running
becuause you were trying to call functions outside of a function. The
code always need some place to start execution. In a normal cpp
program that is the main() function. In a root macro, it will on
loading, attempt to call a function of the same name as the macro
file. You can also call functions defined in the macro after loading
through the ROOT terminal.