Refactor AtHitCluster, AtEvent, and AtPSA (#145)
* HitCluster tracks own position and covariance

AtHits can be directly added to a AtHitCluster. The hit cluster will
update its position, charge, and covariance accordingly using both
charge wieghting and the variance on the position of the hit. Uses a single
pass online algorithm. Details on what is exactly calcualted is in the doxygen.

* Refactor AtPSA so the main logic unit is AtPad -> AtHits.

The managment of summary statistics is now handled by the AtPSA base
class.
Currently ony AtPSAMax is currently refactored

* Refactor AtEvent to store unique_ptr<AtHit> instead of AtHit

Adds in all the refactored PSAs as well.

Verified summary statistics match and AtPRA and AtSampleConsensus are
working as expected.

* Refactor AtHitCluster

Moved extra estiamtes of posisiton and covariance (2-pass algorithms)
to a new class, AtHitClusterFull.

Refactor AtContainerManip into AtData so we don't have a circular
dependence between AtData and AtTools.

Update accessors in AtEvent and deprecate old

Supress compiler warnings we cannot controll