Updates (#176)
* modify to run the d,2He simu, digi and analysis

* check IonDecay and findVertex with more than 2 tracks

* old weighted gaussian hit sampling and other adjustments

* update AtFindVertex

* AtPatternLine 3d fit without q weight

* Ran code linters and updated so everything builds.

Still need to actually review what the code is doing.

* Move S800 specific code into new class for AtTPCIonGenerator

Also moves code for using a gaussian beam spot into a derived class.

* Clean AtPattern

Move other method of drawing a line to AtPattern, and allow for the
max radius of the cylinder to vary so it could work with pAT-TPC as
well.

Also restore optionsl charge weighted fitting

Only draw vertex from AtFindVertex if a boolean is set (AtFindVertex
will seg fault if the pattern for the tracks is anything but an AtPatternLine)

* debug track visu, ionGeneS800

* clang format

* ionGeneS800 detached ata/bta root files

Co-authored-by: Simon <giraud@frib.msu.edu>
Co-authored-by: Adam Anthony <anthonya@nscl.msu.edu>