Proof of concept data subject/observer for a PSA
I think these should be defined in the same header as the PSA
itself. It keeps everything together in a single unit and the
modification of the data of a PSA implies the correspinding
subject/observer also has to change. In principle the internal data of
the PSA could be stored just as a subject that others can access but
that seems unecessary. Maybe in the future (but would be a change that
doesn't effect the public interface at all so could be done
whenever. Would require "ROOTing" the subjects and making them derive
from TObject which I would prefer not to do.