Refactor AtEventManager new to AtViewerManager.
Changed AtEventManagerNew to no longer inherit from
TEveEventManager. AtPadMain should be responsible for the event
managers associated with drawing events as it is the class that
handles creation and updating of the TEveElements used to draw things.

Changed the name of AtEventManagerNew to reflect this difference. It
manages things about the viewer not events either in the TEve event
sense or the AtEvent sense.