Refactor to use subject/observer instead of DrawPad(int)
Anyone can now hook into `fPadNum` in `AtViewerManager` and be
informed whenever the pad we are examining may have changed, either
because it is a different event or a different pad was selected.

Simplifies the need to propogate DrawPad() everywhere especially when
some tabs may have no need up updating on a pad change.