Add pad info to sidebar
Tried listing the electronic ID as a TGLabel but it is costly to remap
the sidepar each time a pad is selected. Probably should replace with
TGNumberEntry to display/select pads on a electronic channel basis.

Probably this should be an optional addition to the sidebar rather
than a permentant addition in the default sidebar.