Add different energy loss models to code and tools to use them. (#194)
Adds a cubic spline class that I modified to track the integral of the spline as well (very little computation to calculate the integral of each spline using Simpson's rule which is also exact for cubics).

Adds a base AtELossModel class for all energy loss models we may want in the future.

Adds AtELossTable for energy loss models from tables. At the moment can load SRIM and LISE++ tables. Verified that energy loss in LISE and range in SRIM match the output of the code using the appropriate table. Because it uses the pre-computed integrals, and a Newton's method-like algorithm for calculating energy losses it should be very fast.

* Add ability to load SRIM tables without editing output

* Add ability to load LISE tables for energy loss