Fix reaction never generating
Because we are setting the event type at the end
of the generator phase (adding particles to the
primary track), we loose the information about
if it is reaction/beam in the stepping phase where
`ProcessHits` is called. To account for this we
use the trackID to determine what the event type
is. Alternatively, we could check
AtVertexPropogator we would just have to invert it

To account for hehavior in `AtTPC2Body` generator
added flag to `AtReactionGenerator` to allow for
any reaction to always run (even if it is a beam-
like event).