Combine detectors into single library
This should help control the many root-level folders.

Adjusted the root dictionary generation to allow for multiple
dictionaries to be added to any target (library). This can speed up
build times for larger libraries if the dictionaries are split up
logically. The dictionaries are each defined by their own linkdefs.
See `AtDetectors` for an example.