Add OpenKF to the AtReconstruction library
Allow additional headers to be added to libraries that are not used in
the generation of the ROOT dictionary. As a result these are not in the
ROOT map, so are not accessible from the ROOT interpreter, but can be
included in other C++ code in the library. This works for OpenKF as I
think we will have classes that wrap these header only classes which
can be easily included in the ROOT map.

Needed to rename util.h to kf_util.h to avoid conflict with tripclust,
which also includes a util.h header.

Removed the main function from tripclust. It interferes with the
generation of any tests that link against the AtReconstruction library.

Add existing OpenKF tests to our test suite

Updated clang-format version to v17.

Version 16 (what we were using) is not available in some Ubuntu
distributions, but v17 is widely available though package managers with
LLVM's official APT repo.