Added container in AtTrack in order to store the experimental values of (range, eLoss).
Added a base class for possible classes that modify a given AtPatternEvent.

Added the task to apply AtPatternModifications to the AtPatternEvents.

Oops.

Added the option to pass AtEvent and AtRawEvent to the AtPatternModification.

Added a virtual method to the base class AtPattern that computes the distance between 2 points along the pattern. Implemented it for the case of AtPatternLine.

Implemented a virtual method in base class AtPattern to obtain the parameter at a given point of the AtPattern.

Modified the AtPSAHitPerTB.

Forgot to change version number.

Implemented an AtPatternModification which extracts the pair of values for the Bragg curve reconstruction and saves them in the AtTracks.

The AtPatternModifications need to be able to write to disk.

Added quick test macros to check if this feature works so far.

Generated the Bragg curve histogram and saved it in a new BraggCurve struct in AtTrack.

Added an AtTab to visualize the Bragg curves of the AtTacks.

Implemented the smoothing of the Bragg curves.